Seed was beign truncated to 0 (size_t is too big)

+++ epiworld.hpp
@@ -7411,21 +7411,21 @@ inline void Model<TSeq>::run_multiple(
         this->seed(seed_);
 
     // Seeds will be reproducible by default
-    std::vector< size_t > seeds_n(nexperiments);
+    std::vector< int > seeds_n(nexperiments);
     #ifdef EPI_DEBUG
     std::fill(
         seeds_n.begin(),
         seeds_n.end(),
         std::floor(
-            runif() * static_cast<double>(std::numeric_limits<size_t>::max())
+            runif() * static_cast<double>(std::numeric_limits<int>::max())
         )
         );
     #else
     for (auto & s : seeds_n)
     {
-        s = static_cast<size_t>(
+        s = static_cast<int>(
             std::floor(
-                runif() * static_cast<double>(std::numeric_limits<size_t>::max())
+                runif() * static_cast<double>(std::numeric_limits<int>::max())
                 )
         );
     }
